// UNpaperWebUI/src/@fott/models/applicationState.ts
export interface ITrainRecordProps {
  accuracies?: object;
  averageAccuracy?: number;
  modelInfo: {
    isComposed?: boolean;
    modelId: string;
    createdDateTime: string;
    modelName: string;
  };
}
// ...
export interface IProviderOptions {
  [key: string]: any;
}
// ...
export interface IProject {
  id: string;
  name: string;
  version: string;
  securityToken: string;
  description?: string;
  tags: ITag[];
  sourceConnection: IConnection;
  assets?: { [index: string]: IAsset };
  lastVisitedAssetId?: string;
  apiUriBase: string;
  apiKey?: string | ISecureString;
  apiVersion?: string;
  folderPath: string;
  trainRecord: ITrainRecordProps;
  recentModelRecords: IRecentModel[];
  predictModelId: string;
}
// ...
export interface ITag {
  name: string;
  color: string;
  type: FieldType;
  format: FieldFormat;
  documentCount?: number;
}
// ...
export interface IConnection {
  id: string;
  name: string;
  description?: string;
  providerType: string;
  providerOptions: IProviderOptions | ISecureString;
}

/**
 * @name - Asset
 * @description - Defines an asset within a project
 * @member id - Unique identifier for asset
 * @member type - Type of asset (Image, Video, etc)
 * @member name - Generated name for asset
 * @member path - Relative path to asset within the underlying data source
 * @member size - Size / dimensions of asset
 * @member format - The asset format (jpg, png, mp4, etc)
 */
export interface IAsset {
  id: string;
  type: AssetType;
  state: AssetState;
  labelingState?: AssetLabelingState;
  name: string;
  path: string;
  size: ISize;
  format?: string;
  timestamp?: number;
  predicted?: boolean;
  ocr?: any;
  isRunningOCR?: boolean;
  isRunningAutoLabeling?: boolean;
  cachedImage?: string;
  mimeType?: string;
  schema?: string;
}
// ...
export interface ISize {
  width: number;
  height: number;
}
// ...
export interface ISecureString {
  encrypted: string;
}
// ...
export interface IRecentModel {
  readonly composedTrainResults?: object;
  readonly accuracies?: object;
  readonly averageAccuracy?: number;
  readonly modelInfo: {
    readonly isComposed: boolean;
    readonly modelId: string;
    readonly createdDateTime: string;
    readonly modelName: string;
  };
}
// ...
export enum AssetType {
  Unknown = 0,
  Image = 1,
  PDF = 5,
  TIFF = 6,
}
// ...
export enum AssetState {
  NotVisited = 0,
  Visited = 1,
  Tagged = 2,
}
// ...
export enum AssetLabelingState {
  ManuallyLabeled = 1,
  Trained = 2,
  AutoLabeled = 3,
  AutoLabeledAndAdjusted = 4,
}
// ...
export enum FieldType {
  String = 'string',
  Number = 'number',
  Date = 'date',
  Time = 'time',
  Integer = 'integer',
  SelectionMark = 'selectionMark',
  Array = 'array',
  Object = 'object',
}

export enum LabelType {
  DrawnRegion = 'region',
}

export enum TableElements {
  rows = 'rows',
  row = 'row',
  columns = 'columns',
  column = 'column',
}

export enum FieldFormat {
  NotSpecified = 'not-specified',
  Currency = 'currency',
  Decimal = 'decimal',
  DecimalCommaSeparated = 'decimal-comma-seperated',
  NoWhiteSpaces = 'no-whitespaces',
  Alphanumeric = 'alphanumeric',
  DMY = 'dmy',
  MDY = 'mdy',
  YMD = 'ymd',
}